requirements.txt
# Core scientific stack
numpy==1.26.4
scipy==1.15.3
pandas==2.3.0
sympy==1.14.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Deep learning
torch==2.7.1
torchvision==0.22.1
torchaudio           # pas dans ta première liste → laissé libre
triton==3.3.1
timm                 # requis pour DETR (pas dans ta première liste → laissé libre)
transformers         # requis pour DETR (pas de version spécifique)
pycocotools          # pour mAP COCO (pas de version spécifique)

# Computer vision
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
pillow==11.2.1
pillow_heif==0.22.0

# Visualization
matplotlib==3.8.4
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
kiwisolver==1.4.8
pyparsing==3.2.3

# Utilities
tqdm==4.67.1
pyyaml==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
python-dotenv==1.1.0
packaging==25.0
platformdirs==4.4.0
filelock==3.18.0
typing_extensions==4.14.0
typing-inspection==0.4.1
six==1.17.0

# Ultralytics & related
ultralytics==8.3.155
ultralytics-thop==2.0.14
roboflow==1.1.65

# HuggingFace / HF ecosystem
huggingface-hub==0.34.4
hf-xet==1.1.9

# Gradio / FastAPI stack
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.3
gradio==5.44.1
gradio_client==1.12.1
anyio==4.10.0
sniffio==1.3.1
aiofiles==24.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
orjson==3.11.3
pydantic==2.11.7
pydantic_core==2.33.2
annotated-types==0.7.0
python-multipart==0.0.20

# Visualization / notebooks / syntax highlight
Pygments==2.19.2
markdown-it-py==4.0.0
mdurl==0.1.2
rich==14.1.0

# Logging / monitoring
wandb==0.21.3
sentry-sdk==2.37.1

# Audio / media
pydub==0.25.1
ffmpy==0.6.1

# Git / versioning
GitPython==3.1.45
gitdb==4.0.12
smmap==5.0.2

# NVIDIA stack (si GPU avec CUDA)
nvidia-ml-py==12.575.51
nvidia-pyindex
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
pynvml==12.0.0

# Miscellaneous
certifi==2025.6.15
charset-normalizer==3.4.2
urllib3==2.4.0
Brotli==1.1.0
idna==3.7
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
typer==0.17.3
ruff==0.12.12
psutil==7.0.0
groovy==0.1.2
mpmath==1.3.0